+++ src/config/Config.ts
@@ -5,7 +5,7 @@ import RNConfig from 'react-native-config';
 import DeviceInfo from 'react-native-device-info';
 import { DevMode } from './DevMode';
 
-export const CODEPUSH_VERSION = '4.31.2';
+export const CODEPUSH_VERSION = '4.31.3';
 
 export const IS_IOS = Platform.OS === 'ios';
 export const IS_IPAD = (Platform as PlatformIOSStatic).isPad;
@@ -170,6 +170,8 @@ export const MINDS_DEEPLINK = [
     },
   ],
   ['supermind/:guid', 'Supermind', 'navigate'],
+  ['boost/boost-console', 'BoostConsole', 'navigate'],
+  ['boost/console/newsfeed/history', 'BoostConsole', 'navigate'],
 ];
 
 // IF TRUE COMMENT THE SMS PERMISSIONS IN ANDROID MANIFEST TOO!!!