+++ src/modules/boost/boost-composer/screens/BoostComposer.tsx
@@ -1,9 +1,9 @@
 import { observer } from 'mobx-react';
 import React from 'react';
-import { confirm } from '../../../common/components/Confirm';
-import FitScrollView from '../../../common/components/FitScrollView';
-import Slider from '../../../common/components/Slider';
-import TopbarTabbar from '../../../common/components/topbar-tabbar/TopbarTabbar';
+import { confirm } from '~/common/components/Confirm';
+import FitScrollView from '~/common/components/FitScrollView';
+import Slider from '~/common/components/Slider';
+import TopbarTabbar from '~/common/components/topbar-tabbar/TopbarTabbar';
 import {
   B1,
   B2,
@@ -14,17 +14,19 @@ import {
   HairlineRow,
   Screen,
   ScreenHeader,
-} from '../../../common/ui';
-import ThemedStyles from '../../../styles/ThemedStyles';
+} from '~/common/ui';
+import ThemedStyles from '~/styles/ThemedStyles';
+import { useTranslation } from '../../locales';
 import { IPaymentType, useBoostStore } from '../boost.store';
-import { useTranslation } from '../locales';
 import { BoostStackScreenProps } from '../navigator';
+import useBoostInsights from '../../hooks/useBoostInsights';
 
 type BoostComposerScreenProps = BoostStackScreenProps<'BoostComposer'>;
 
 function BoostComposerScreen({ navigation }: BoostComposerScreenProps) {
   const { t } = useTranslation();
   const boostStore = useBoostStore();
+  const { insights } = useBoostInsights(boostStore);
 
   const textMapping = {
     cash: {
@@ -117,7 +119,11 @@ function BoostComposerScreen({ navigation }: BoostComposerScreenProps) {
             {t('Total Spend')}
           </B1>
 
-          <H2 bottom="S">{t('Unknown')}</H2>
+          <H2 bottom="S">
+            {insights
+              ? `${insights?.views?.low?.toLocaleString()} - ${insights?.views?.high?.toLocaleString()}`
+              : t('Unknown')}
+          </H2>
           <B1 color="secondary">{t('Estimated reach')}</B1>
         </Column>