+++ src/navigation/NavigationStack.tsx
@@ -225,6 +225,18 @@ const AppStack = observer(() => {
           getComponent={() => require('~/supermind/SupermindScreen').default}
           options={hideHeader}
         />
+        <AppStackNav.Screen
+          name="Referrals"
+          getComponent={() => require('~/referral/ReferralsScreen').default}
+          options={{
+            title: 'Invite Friends',
+          }}
+        />
+        <AppStackNav.Screen
+          name="BoostConsole"
+          getComponent={() => require('modules/boost').BoostConsoleScreen}
+          options={hideHeader}
+        />
       </AppStackNav.Navigator>
     </>
   );
@@ -426,7 +438,7 @@ const RootStack = observer(function () {
             />
             <RootStackNav.Screen
               name="BoostScreen"
-              getComponent={() => require('~/boost/v2/BoostScreen').default}
+              getComponent={() => require('~/boost/legacy/BoostScreen').default}
               options={modalOptions}
             />
             <RootStackNav.Screen
@@ -486,7 +498,7 @@ const RootStack = observer(function () {
             />
             <RootStackNav.Screen
               name="BoostScreenV2"
-              getComponent={() => require('~/modules/boost').BoostStack}
+              getComponent={() => require('modules/boost').BoostComposerStack}
               options={modalOptions}
             />
           </>