+++ src/navigation/NewsfeedStack.tsx
@@ -50,7 +50,7 @@ export default function () {
       />
       <NewsfeedStack.Screen
         name="BoostScreenV2"
-        getComponent={() => require('~/modules/boost').BoostStack}
+        getComponent={() => require('modules/boost').BoostComposerStack}
         options={hideHeader}
       />
     </NewsfeedStack.Navigator>