+++ src/newsfeed/SeeLatestButton.tsx
@@ -1,41 +1,36 @@
 import { observer } from 'mobx-react';
 import React, { useEffect, useRef, useState } from 'react';
-import Animated, { FadeInUp, useAnimatedStyle } from 'react-native-reanimated';
-import { useFeedListContext } from '~/common/components/FeedListSticky';
+import Animated, {
+  FadeInUp,
+  useAnimatedStyle,
+  useDerivedValue,
+  withTiming,
+} from 'react-native-reanimated';
 import i18nService from '~/common/services/i18n.service';
 import { Button, Icon } from '~/common/ui';
 import { IS_IOS, NEWSFEED_NEW_POST_POLL_INTERVAL } from '~/config/Config';
+import {
+  ScrollDirection,
+  useScrollContext,
+} from '~/common/contexts/scroll.context';
 import useApiFetch from '../common/hooks/useApiFetch';
+import { Timeout } from '../types/Common';
 
 interface SeeLatestButtonProps {
   onPress?: () => Promise<void>;
   countEndpoint: string;
 }
 
-const additionalTop = IS_IOS ? 160 : 150;
+const additionalTop = 163;
+const DISMISS_TIMEOUT = 5000;
+const TOP_HIDDEN_OFFSET = -500;
 
 /**
  * A prompt that appears in a feed and shows how many new posts are there
  */
 const SeeLatestButton = ({ onPress, countEndpoint }: SeeLatestButtonProps) => {
-  const context = useFeedListContext();
   const { count, resetCount } = useWatchForUpdates(countEndpoint);
-  const scrollY = context?.scrollY;
-  const style = useAnimatedStyle(() => {
-    const margin = additionalTop - (IS_IOS ? 70 : 60);
-    let translateY = scrollY
-      ? scrollY.value < margin
-        ? scrollY.value
-        : margin
-      : 0;
-
-    return {
-      top: additionalTop - translateY,
-      position: 'absolute',
-      left: 0,
-      right: 0,
-    };
-  });
+  const style = useSeeLatestStyle(count, 170);
   const handleOnPress = () => onPress?.().then(() => resetCount());
 
   if (!count) {
@@ -64,6 +59,62 @@ const SeeLatestButton = ({ onPress, countEndpoint }: SeeLatestButtonProps) => {
 };
 
 export default observer(SeeLatestButton);
+
+export const useSeeLatestStyle = (count: number, top?: number) => {
+  const context = useScrollContext();
+  const {
+    scrollY = { value: 0 },
+    translationY = { value: 0 },
+    scrollDirection = { value: 0 },
+  } = context ?? {};
+  const topDistance = top || additionalTop;
+  const [dismissible, setDismissible] = useState(false);
+  const countAvailable = !!count;
+  const timeOutRef = useRef<Timeout>();
+
+  const dismissed = useDerivedValue(() => {
+    return dismissible && scrollDirection.value === ScrollDirection.down;
+  }, [scrollDirection, dismissible]);
+
+  useEffect(() => {
+    if (countAvailable) {
+      timeOutRef.current = setTimeout(() => {
+        setDismissible(true);
+      }, DISMISS_TIMEOUT);
+    } else {
+      setDismissible(false);
+    }
+
+    return () => {
+      if (timeOutRef.current) {
+        clearTimeout(timeOutRef.current);
+      }
+    };
+  }, [countAvailable]);
+
+  return useAnimatedStyle(() => {
+    const margin = topDistance - (IS_IOS ? 70 : 65);
+
+    let translateY = scrollY
+      ? scrollY.value < margin
+        ? scrollY.value
+        : margin + translationY.value
+      : 0;
+
+    return {
+      top: withTiming(
+        dismissed.value ? TOP_HIDDEN_OFFSET : topDistance - translateY,
+        {
+          duration: 300,
+        },
+      ),
+      position: 'absolute',
+      left: 0,
+      right: 0,
+      zIndex: 0,
+    };
+  }, [dismissed]);
+};
 
 const useWatchForUpdates = (countEndpoint: string) => {
   const previousCount = useRef<number>();