+++ src/supermind/SupermindConsoleScreen.tsx
@@ -25,6 +25,15 @@ import SupermindConsoleFeedFilter, {
 } from './SupermindConsoleFeedFilter';
 import SupermindRequest from './SupermindRequest';
 import SupermindRequestModel from './SupermindRequestModel';
+import {
+  useAnimatedScrollHandler,
+  useSharedValue,
+} from 'react-native-reanimated';
+import Animated from 'react-native-reanimated';
+import {
+  ScrollContext,
+  ScrollDirection,
+} from '~/common/contexts/scroll.context';
 
 type TabModeType = 'inbound' | 'outbound';
 type SupermindConsoleScreenRouteProp = RouteProp<
@@ -36,6 +45,7 @@ type SupermindConsoleScreenNavigationProp = StackNavigationProp<
   'SupermindConsole'
 >;
 
+const MIN_SCROLL_THRESHOLD = 5;
 const filterValues: Record<SupermindFilterType, string> = {
   all: '',
   pending: '1',
@@ -66,6 +76,8 @@ function SupermindConsoleScreen({
   const listRef = React.useRef<any>(null);
   const [onboarding, dismissOnboarding] = useSupermindOnboarding('producer');
   const isStripeConnectFeatureOn = useIsFeatureOn('mob-stripe-connect-4587');
+  const scrollDirection = useSharedValue(0);
+  const scrollY = useSharedValue(0);
 
   const scrollToTopAndRefresh = () => {
     listRef.current?.scrollToTop();
@@ -80,6 +92,20 @@ function SupermindConsoleScreen({
       setFilter('all');
     }
   };
+
+  const scrollHandler = useAnimatedScrollHandler({
+    onScroll: event => {
+      if (
+        Math.abs(event.contentOffset.y - scrollY.value) > MIN_SCROLL_THRESHOLD
+      ) {
+        scrollDirection.value =
+          event.contentOffset.y > scrollY.value
+            ? ScrollDirection.down
+            : ScrollDirection.up;
+      }
+      scrollY.value = event.contentOffset.y;
+    },
+  });
 
   const filterParam = filterValues[filter]
     ? `?status=${filterValues[filter]}`
@@ -118,6 +144,7 @@ function SupermindConsoleScreen({
       />
       <OffsetList
         ref={listRef}
+        ListComponent={Animated.FlatList}
         header={
           <>
             <TopbarTabbar
@@ -153,14 +180,23 @@ function SupermindConsoleScreen({
           mode === 'inbound' ? renderSupermindInbound : renderSupermindOutbound
         }
         endpointData=""
+        onScroll={scrollHandler}
       />
 
-      <SeeLatestButton
-        countEndpoint={`api/v3/supermind/${
-          mode === 'inbound' ? 'inbox' : 'outbox'
-        }/count${filterParam}`}
-        onPress={scrollToTopAndRefresh}
-      />
+      <ScrollContext.Provider
+        value={{
+          scrollDirection,
+          translationY: { value: -60 },
+          headerHeight: 0,
+          scrollY,
+        }}>
+        <SeeLatestButton
+          countEndpoint={`api/v3/supermind/${
+            mode === 'inbound' ? 'inbox' : 'outbox'
+          }/count${filterParam}`}
+          onPress={scrollToTopAndRefresh}
+        />
+      </ScrollContext.Provider>
 
       <AnimatePresence>
         {onboarding && (