+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 Minds Mobile
 
+## 4.31.4 (2023-01-23)
+
+### changed (1 change)
+
+- [(feat) add boost insight estimated reach to boost composer](minds/mobile-native@4a15f2d34f098e4afcd2b11cfd97050620bb6ec2) ([merge request](minds/mobile-native!2058))
+
+## 4.31.3 - 2023-01-18
+
+### Changed
+
+- Allow boosting all posts
+- Improved layout for tablets
+- Boost v3 notification
+- Pending supermind notice navigation fixes
+- Fix setup channel in-feed notice not shown
+- in-feed notice invite your friends
+- Base design module created using Tamagui
+- Fix different subscribers list between web & mobile
+- Fix issues on the interaction bottom sheet
+- Moved boosted label to the top of the posts
+
 ## 4.31.2 - 2023-01-06
 
 ### Changed